header, comp
In comp, a function for converting from a path to a file name with .lz78 extension has been implemented; 
In header, scanf has been substituted with fread, and fprintf with fwrite. This way if may be less efficient, but it works.
Minor changes has been done here and there (comments, print etc)